+++ include/ambition/Ambition_building.hh
@@ -160,6 +160,9 @@ public:
   Coordinates::Point getRallyLocation(
   ) const;
 
+  bool mainMenuActive(
+  );
+
   void migrate7kaaQueue(
   );